daemon: move utxo `jsonrpc_` methods to an individual module
