+++ lbry/extras/daemon/daemon.py
@@ -231,6 +231,7 @@ from lbry.extras.daemon.daemon_collection import Daemon_collection
 from lbry.extras.daemon.daemon_support import Daemon_support
 from lbry.extras.daemon.daemon_transaction import Daemon_transaction
 from lbry.extras.daemon.daemon_txo import Daemon_txo
+from lbry.extras.daemon.daemon_utxo import Daemon_utxo
 
 HISTOGRAM_BUCKETS = (
     .005, .01, .025, .05, .075, .1, .25, .5, .75, 1.0, 2.5, 5.0, 7.5, 10.0, 15.0, 20.0, 30.0, 60.0, float('inf')
@@ -241,7 +242,7 @@ class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
              Daemon_account, Daemon_sync, Daemon_address, Daemon_file,
              Daemon_purchase, Daemon_claim, Daemon_channel,
              Daemon_stream, Daemon_collection, Daemon_support,
-             Daemon_transaction, Daemon_txo):
+             Daemon_transaction, Daemon_txo, Daemon_utxo):
     """
     LBRYnet daemon, a jsonrpc interface to lbry functions
     """
@@ -882,53 +883,8 @@ class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
     UTXO_DOC = """
     Unspent transaction management.
     """
-
-    @requires(WALLET_COMPONENT)
-    def jsonrpc_utxo_list(self, *args, **kwargs):
-        """
-        List unspent transaction outputs
-
-        Usage:
-            utxo_list [<account_id> | --account_id=<account_id>] [--wallet_id=<wallet_id>]
-                      [--page=<page>] [--page_size=<page_size>]
-
-        Options:
-            --account_id=<account_id>  : (str) id of the account to query
-            --wallet_id=<wallet_id>    : (str) restrict results to specific wallet
-            --page=<page>              : (int) page to return during paginating
-            --page_size=<page_size>    : (int) number of items on page during pagination
-
-        Returns: {Paginated[Output]}
-        """
-        kwargs['type'] = ['other', 'purchase']
-        kwargs['is_not_spent'] = True
-        return self.jsonrpc_txo_list(*args, **kwargs)
-
-    @requires(WALLET_COMPONENT)
-    async def jsonrpc_utxo_release(self, account_id=None, wallet_id=None):
-        """
-        When spending a UTXO it is locally locked to prevent double spends;
-        occasionally this can result in a UTXO being locked which ultimately
-        did not get spent (failed to broadcast, spend transaction was not
-        accepted by blockchain node, etc). This command releases the lock
-        on all UTXOs in your account.
-
-        Usage:
-            utxo_release [<account_id> | --account_id=<account_id>] [--wallet_id=<wallet_id>]
-
-        Options:
-            --account_id=<account_id> : (str) id of the account to query
-            --wallet_id=<wallet_id>   : (str) restrict operation to specific wallet
-
-        Returns:
-            None
-        """
-        wallet = self.wallet_manager.get_wallet_or_default(wallet_id)
-        if account_id is not None:
-            await wallet.get_account_or_error(account_id).release_all_outputs()
-        else:
-            for account in wallet.accounts:
-                await account.release_all_outputs()
+    # jsonrpc_utxo_list
+    # jsonrpc_utxo_release
 
     BLOB_DOC = """
     Blob management.
